filtering out unneeded dividers, removing hidden elements blank space
