+++ src/containers/SettingsEditor/FormTemplates/FieldTemplate.tsx
@@ -8,6 +8,8 @@ import copyToClipboard from '@helpers/copyToClipboard'
 import { $Any } from '@types'
 import { FieldTemplateProps } from '@rjsf/utils'
 import { CSS } from 'styled-components/dist/types'
+import { matchesFilterKeys } from './helpers'
+import clsx from 'clsx'
 
 const arrayStartsWith = (arr1: $Any, arr2: $Any) => {
   // return true, if first array starts with second array
@@ -36,10 +38,13 @@ function FieldTemplate(props: FieldTemplateProps) {
   }
 
   const divider = useMemo(() => {
-    if (props.schema.section)
-      return <Divider>{props.schema.section !== '---' && props.schema.section}</Divider>
-    else return <></>
-  }, [props.schema.section])
+    const matches = matchesFilterKeys(props.formContext.filterKeys, props.id)
+    if (props.schema.section && matches) {
+      return <Divider> {props.schema.section !== '---' && props.schema.section} </Divider>
+    }
+
+    return <></>
+  }, [props.schema.section, props.schema.filterKeys])
 
   // Object fields
 
@@ -129,22 +134,34 @@ function FieldTemplate(props: FieldTemplateProps) {
 
     if (!classes.includes('obj-override-edit')) classes.push(`obj-override-${overrideLevel}`)
 
+    const matches = matchesFilterKeys(props.formContext.filterKeys, props.id)
+
     return (
-      <SettingsPanel
-        objId={props.id}
-        title={props.schema.title}
-        description={props.schema.description}
-        className={classes.join(' ')}
-        onClick={() => {
-          if (props.formContext.onSetBreadcrumbs && path) props.formContext.onSetBreadcrumbs(path)
+      <div
+        data-schema-id={props.id}
+        data-hidden={matches ? 'false' : 'true'}
+        style={{
+          visibility: matches ? 'visible' : 'hidden',
+          position: matches ? 'relative' : 'absolute',
+          height: matches ? 'auto' : 0,
         }}
-        onContextMenu={onContextMenu}
-        currentId={props.formContext.currentId}
-        layout={undefined}
-        enabledToggler={undefined}
       >
-        {props.children}
-      </SettingsPanel>
+        <SettingsPanel
+          objId={props.id}
+          title={props.schema.title}
+          description={props.schema.description}
+          className={classes.join(' ')}
+          onClick={() => {
+            if (props.formContext.onSetBreadcrumbs && path) props.formContext.onSetBreadcrumbs(path)
+          }}
+          onContextMenu={onContextMenu}
+          currentId={props.formContext.currentId}
+          layout={undefined}
+          enabledToggler={undefined}
+        >
+          {props.children}
+        </SettingsPanel>
+      </div>
     )
   }
